+++ src/createWebpackConfig.js
@@ -433,9 +433,9 @@ export function createExtraRules(
 }
 
 /**
- * Plugin for HtmlPlugin which inlines content for an extracted Webpack
- * manifest into the HTML page in a <script> tag before other emitted asssets
- * are injected by HtmlPlugin itself.
+ * Plugin for HtmlPlugin which inlines content for an extracted Webpack manifest
+ * into the HTML in a <script> tag before other emitted asssets are injected by
+ * HtmlPlugin itself.
  */
 function injectManifestPlugin() {
   this.plugin('compilation', (compilation) => {
@@ -448,8 +448,8 @@ function injectManifestPlugin() {
             /^(\s*)<\/body>/m,
             `$1<script>${children[0]._value}</script>\n$1</body>`
           )
-          // Remove the manifest from HtmlPlugin's assets to
-          // prevent a <script> tag being created for it.
+          // Remove the manifest from HtmlPlugin's assets to prevent a <script>
+          // tag being created for it.
           var manifestIndex = data.assets.js.indexOf(data.assets.publicPath + key)
           data.assets.js.splice(manifestIndex, 1)
           delete data.assets.chunks.manifest
@@ -595,7 +595,8 @@ export function createPlugins(
   }
 
   // Automatically install missing npm dependencies and add them to package.json
-  // Must be enabled with an --install or --auto-install flag
+  // if present.
+  // Must be enabled with an --install or --auto-install flag.
   if (buildConfig.autoInstall) {
     plugins.push(new NpmInstallPlugin({
       peerDependencies: false,