+++ src/webpackServer.js
@@ -23,7 +23,9 @@ function getServerPort(args, cb) {
     // Support use of --force to avoid interactive prompt
     if (args.force) return cb(null, suggestedPort)
 
-    clearConsole()
+    if (args.clear !== false && args.clearConsole !== false) {
+      clearConsole()
+    }
     console.log(yellow(`Something is already running on port ${intendedPort}.`))
     console.log()
     inquirer.prompt([
@@ -69,7 +71,8 @@ export default function webpackServer(args, buildConfig, cb) {
 
     if (!('status' in buildConfig.plugins)) {
       buildConfig.plugins.status = {
-        message: `The app is running at http://${args.host || 'localhost'}:${port}/`,
+        disableClearConsole: args.clear !== false && args.clearConsole !== false,
+        successMessage: `The app is running at http://${args.host || 'localhost'}:${port}/`,
       }
     }